move errors.ts and events.ts to lib folder
