+++ src/lib/errors.ts
